Update CMSC495_Home.html
Adding the login and signup overlays.  Changed the footer to have our names in the project contributors.  Changed the formatting for all of the pages to have a max-width since objects were way too big if the website was viewed in full screen.  Added some SWEET animations for the overlays.